requirements.txt
numpy==1.22.4
pygame==2.1.2
tqdm==4.64.0
scikit-learn==1.1.2
matplotlib==3.5.3
pandas==1.4.4
python-stopwatch